+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Bug fixes
 * CONRAT-34249: Fix ref type for first pv event after app-to-app tracking.
 * CONRAT-35311: Refuse zero and negative values for account and application identifiers when adding a duplicate account.
+* CONRAT-34265: Fix duplicate `_rem_launch` event on app launch for apps with `SceneDelegate`.
 
 ## 10.4.0 (2024-11-14)